move create_dir function to a separate utils module
